Fix InMemory Combiner
The mapper function is called once per line. In the old implementation we created one HashMap per line which does not make much sense.

This implementation uses the setup() and cleanup() methods to create the HashMap for the mapper before the first line of text is processed. After all lines are read, all pairs are emitted in the close() method.